backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
typer==0.12.3
pydantic==2.7.1
aiosqlite==0.20.0
# Actian VectorAI DB Python client — install manually from the .whl in the repo:
#   pip install actiancortex-0.1.0b1-py3-none-any.whl
